feat(ui): applied view bindings
