+++ bottom_navigation/app/src/main/java/com/example/bottom_navigation/ui/DashboardFragment.kt
@@ -1,4 +1,4 @@
-package com.example.bottom_navigation.ui.dashboard
+package com.example.bottom_navigation.ui
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -6,26 +6,23 @@ import android.view.View
 import android.view.ViewGroup
 import android.widget.TextView
 import androidx.fragment.app.Fragment
+import androidx.fragment.app.viewModels
 import androidx.lifecycle.Observer
 import androidx.lifecycle.ViewModelProvider
 import com.example.bottom_navigation.R
+import com.example.bottom_navigation.databinding.FragmentDashboardBinding
 
 class DashboardFragment : Fragment() {
 
-    private lateinit var dashboardViewModel: DashboardViewModel
+    private lateinit var binding : FragmentDashboardBinding
 
     override fun onCreateView(
             inflater: LayoutInflater,
             container: ViewGroup?,
             savedInstanceState: Bundle?
     ): View? {
-        dashboardViewModel =
-                ViewModelProvider(this).get(DashboardViewModel::class.java)
-        val root = inflater.inflate(R.layout.fragment_dashboard, container, false)
-        val textView: TextView = root.findViewById(R.id.text_dashboard)
-        dashboardViewModel.text.observe(viewLifecycleOwner, Observer {
-            textView.text = it
-        })
-        return root
+        binding = FragmentDashboardBinding.inflate(layoutInflater, container, false)
+        binding.textDashboard.text = "This is dashboard Fragment"
+        return binding.root
     }
 }
