+++ bottom_navigation/app/src/main/java/com/example/bottom_navigation/ui/HomeFragment.kt
@@ -1,0 +1,23 @@
+package com.example.bottom_navigation.ui
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.fragment.app.Fragment
+import com.example.bottom_navigation.databinding.FragmentHomeBinding
+
+class HomeFragment : Fragment() {
+
+    private lateinit var binding: FragmentHomeBinding
+
+    override fun onCreateView(
+        inflater: LayoutInflater,
+        container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        binding = FragmentHomeBinding.inflate(layoutInflater, container, false)
+        binding.textHome.text = "This is home Fragment"
+        return binding.root
+    }
+}
