+++ bottom_navigation/app/src/main/java/com/example/bottom_navigation/ui/NotificationsFragment.kt
@@ -1,0 +1,23 @@
+package com.example.bottom_navigation.ui
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.fragment.app.Fragment
+import com.example.bottom_navigation.databinding.FragmentNotificationsBinding
+
+class NotificationsFragment : Fragment() {
+
+    private lateinit var binding: FragmentNotificationsBinding
+
+    override fun onCreateView(
+        inflater: LayoutInflater,
+        container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        binding = FragmentNotificationsBinding.inflate(layoutInflater, container, false)
+        binding.textNotifications.text = "This is notifications Fragment"
+        return binding.root
+    }
+}
